Attach the certificate to a renewal confirmation email
