+++ spec/cassettes/notify_upper_tier_renewal_confirmation_sends_an_email.yml
@@ -8,7 +8,7 @@ http_interactions:
       string: '{"email_address":"foo@example.com","template_id":"6d54a9bc-9b62-4d93-a40a-d06d04ed58ca","personalisation":{"reg_identifier":"CBDU1","registration_type":"carrier,
         broker and dealer","first_name":"Jane","last_name":"Doe","phone_number":"03708
         506506","registered_address":"42\r\nFoo Gardens\r\nBaz City\r\nFA1 1KE","date_activated":"30
-        April 2021"}}'
+        April 2021","link_to_file":"My certificate"}}'
     headers:
       User-Agent:
       - NOTIFY-API-RUBY-CLIENT/5.3.0
@@ -34,43 +34,44 @@ http_interactions:
       Content-Type:
       - application/json
       Date:
-      - Fri, 30 Apr 2021 14:15:22 GMT
+      - Fri, 30 Apr 2021 14:36:00 GMT
       Server:
       - nginx
       Strict-Transport-Security:
       - max-age=31536000; includeSubdomains
       X-B3-Spanid:
-      - 11ec306016e3301a
+      - 137b0f5506554061
       X-B3-Traceid:
-      - 11ec306016e3301a
+      - 137b0f5506554061
       X-Vcap-Request-Id:
-      - 5bab280a-9fe0-4378-65a7-bedc553baa31
+      - c93bf6ba-e766-4619-4ca2-11ac43f77bba
       Content-Length:
-      - '1956'
+      - '2008'
       Connection:
       - keep-alive
     body:
       encoding: UTF-8
       string: '{"content":{"body":"#Your registration as an upper tier carrier, broker
-        and dealer has been renewed\r\n\r\nYour registration number is still\r\n\r\n#CBDU1\r\n\r\nBeing
-        an an upper tier waste carrier, broker and dealer means that:\r\n\r\n*you
-        need to renew your registration every 3 years (we\u2019ll send you a reminder)\r\n*you\u2019ll
-        need to pay a registration charge each time you renew\r\n\r\nThere are some
-        simple rules you must follow:\r\n\r\n*if you give your waste to another person
-        or business you must check they are properly authorised to accept it, e.g.
-        as a permitted site or a registered waste carrier\r\n*make sure the correct
-        documentation is completed for each transfer of waste and that it correctly
-        describes the waste\r\n*make sure any waste is safely handled and stored\r\n*minimise
-        the environmental impact of waste by prioritising waste prevention, reuse,
-        recycling and recovery over disposal\r\n\r\n#Updating your details\r\nIf any
-        of the details you\u2019ve given us change, you must update them within 28
-        days by contacting the Environment Agency.\r\n\r\nJane Doe\r\n\r\n42\r\nFoo
-        Gardens\r\nBaz City\r\nFA1 1KE\r\n\r\nTelephone number: 03708 506506\r\n\r\nRenewed
-        on: 30 April 2021\r\n\r\nIf you have enquiries please email nccc-carrierbroker@environment-agency.gov.uk
+        and dealer has been renewed\r\n\r\nYour registration number is still\r\n\r\n#CBDU1\r\n\r\nDownload
+        your certificate at: My certificate\r\n\r\nBeing an an upper tier waste carrier,
+        broker and dealer means that:\r\n\r\n*you need to renew your registration
+        every 3 years (we\u2019ll send you a reminder)\r\n*you\u2019ll need to pay
+        a registration charge each time you renew\r\n\r\nThere are some simple rules
+        you must follow:\r\n\r\n*if you give your waste to another person or business
+        you must check they are properly authorised to accept it, e.g. as a permitted
+        site or a registered waste carrier\r\n*make sure the correct documentation
+        is completed for each transfer of waste and that it correctly describes the
+        waste\r\n*make sure any waste is safely handled and stored\r\n*minimise the
+        environmental impact of waste by prioritising waste prevention, reuse, recycling
+        and recovery over disposal\r\n\r\n#Updating your details\r\nIf any of the
+        details you\u2019ve given us change, you must update them within 28 days by
+        contacting the Environment Agency.\r\n\r\nJane Doe\r\n\r\n42\r\nFoo Gardens\r\nBaz
+        City\r\nFA1 1KE\r\n\r\nTelephone number: 03708 506506\r\n\r\nRenewed on: 30
+        April 2021\r\n\r\nIf you have enquiries please email nccc-carrierbroker@environment-agency.gov.uk
         or contact the Environment Agency helpline: 03708 506506\r\n\r\nThis is an
         automated email, please do not reply","from_email":"waste.carriers.registration.service@notifications.service.gov.uk","subject":"Your
-        waste carriers registration CBDU1 has been renewed"},"id":"fb69a4ab-95cc-4af0-b407-0b8cf4a97dce","reference":null,"scheduled_for":null,"template":{"id":"6d54a9bc-9b62-4d93-a40a-d06d04ed58ca","uri":"https://api.notifications.service.gov.uk/services/25cb6b94-8ce7-485b-918a-559f3b18f69c/templates/6d54a9bc-9b62-4d93-a40a-d06d04ed58ca","version":4},"uri":"https://api.notifications.service.gov.uk/v2/notifications/fb69a4ab-95cc-4af0-b407-0b8cf4a97dce"}
+        waste carriers registration CBDU1 has been renewed"},"id":"c2332345-1093-4c47-9b8b-f46a4f249f5d","reference":null,"scheduled_for":null,"template":{"id":"6d54a9bc-9b62-4d93-a40a-d06d04ed58ca","uri":"https://api.notifications.service.gov.uk/services/25cb6b94-8ce7-485b-918a-559f3b18f69c/templates/6d54a9bc-9b62-4d93-a40a-d06d04ed58ca","version":5},"uri":"https://api.notifications.service.gov.uk/v2/notifications/c2332345-1093-4c47-9b8b-f46a4f249f5d"}
 
 '
-  recorded_at: Fri, 30 Apr 2021 14:15:22 GMT
+  recorded_at: Fri, 30 Apr 2021 14:36:00 GMT
 recorded_with: VCR 6.0.0
